CrewAI_agent/CrewAI_Competitive_Insights_Flask_Cloud_Deployment/requirements.txt
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
humanize==4.10.0
ibis-framework==8.0.0
idna==3.10
importlib_metadata==7.2.1
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.19.1
jellyfish==1.1.0
Jinja2==3.0.2
Flask==1.1.2
Werkzeug==0.16.1
flask-cors==5.0.1
requests==2.32.3
python-dotenv==1.0.1